Progress エンティティを course.entity.go から progress.entity.go に切り出す
Entire-Checkpoint: 09fd8aad5628
